Move all numbered bookmarks bindings to settings
In keybindings, vscode see the captures the first key of the chord
regardless of what other vim commands might be in progress, e.g. rm,
which should replace the character under the cursor with m, instead gets
interpretted as the first key in a chord. Allowing the plugin to handle
the chords instead results in the correct behavior